+++ packages/flowerbase/CHANGELOG.md
@@ -1,3 +1,26 @@
+## 1.8.3 (2026-04-02)
+
+
+### 🩹 Fixes
+
+- updateOne options ([5b16467](https://github.com/flowerforce/flowerbase/commit/5b16467))
+
+## 1.8.2 (2026-03-31)
+
+
+### 🩹 Fixes
+
+- apply_when ([f2bcef2](https://github.com/flowerforce/flowerbase/commit/f2bcef2))
+
+- forwarded host ([#52](https://github.com/flowerforce/flowerbase/pull/52))
+
+## 1.8.1 (2026-03-27)
+
+
+### 🩹 Fixes
+
+- add config FUNCTION_CALL_BODY_LIMIT ([5f6ec70](https://github.com/flowerforce/flowerbase/commit/5f6ec70))
+
 ## 1.8.0 (2026-03-20)